Improved VIF parsing and added support to load models from a chunk file.
Also added support for reading DPL models. Amazing how simple that was!

Now I just need to figure out what to do with this information...

Note:
I added in a little "interactive mode" for fun, just add "--" to the end
of your program arguments and press "space" to load the next model in!