Complete overhaul of vehicle hierarchy stuff. Could use a rewrite...
Not too happy with the way the new code came out,
but then again I didn't know much of the spec when I first wrote it.

Rewriting the code required modifications to the VPK exporter.
Since I'm not sure if I broke anything, I bumped the version up to 7. :P