Updated the older version of Zartex after nearly 2 years.
My curiosity got the better of me, so I decided to check out what was
broken about Zartex. It turns out I had gotten it semi-working in 2015,
but never progressed very far with it.

A lot of the code itself is nearly 4 years old, so it's quite messy!

I really did this just to be able to look at mission scripts again,
and I have no guarantee that editing stuff has any effect.

From this point forward, Zartex will now be included in the
AppVeyor artifacts. As much as I'm ashamed of the poor design and
buggy code, I'm sure there's plenty of people out there who've always
wanted to use it, even if it doesn't do much...

With all of that said:
*** USE AT YOUR OWN RISK! ***