Added a rudimentary model package converter/recompiler.
Currently supports converting Driver: Parallel Lines car models into
the format used in Driv3r. Models already in the Driv3r format will
just be recompiled to use triangle lists, essentially.

This feature is not yet fully implemented, and as such, is pretty
much barely holding together with some toothpicks and glue.

While not ready for the end-user quite yet, you can convert models
by selecting the desired model package and pressing 'X' while in
the Viewport tab. That's only one part, however...

More editing capabilities will need to be added to Antilli in order
for this to actually work, because lots of hand-tweaking is still
necessary; this will definitely be on the roadmap going forward.

There is currently no way to replace models other than to use the
Chunk Editor to extract the new model chunk, and replace an existing
model chunk in another. Yes, it's complicated, and it sucks. Sorry!